restrict height of `TextView` in "shutter" overlay
In a grid view having many columns,
this text can wrap many lines,
which expands the height of the `PlayerView`,
and height does not snap back after the "shutter" overlay is removed.

Consequently, the overall height of this text
needs to be restricted to a sane value for the final row of videos.